Add timeout option to client
There are some places where we make use of varying timeouts when
making HTTP requests to Seed services.

There's no way to timeout a session - the requests devs are very
sure this isn't something they want to add because semantically
a session can't be timed out.

The recommend way to do this is to subclass HTTPAdapter. I've pinned
a specific version of requests as we now depend on an internal bit of
it.